move gpu_stats into components
