+++ lact-gui/src/app/components/gpu_stats_section.css
