+++ lact-gui/src/app/pages/oc_page.rs
@@ -1,7 +1,7 @@
 mod clocks_frame;
 mod vf_curve;
 
-use crate::app::gpu_stats_section::{
+use crate::app::components::gpu_stats_section::{
     GpuStat, GpuStatsSection, GpuStatsSectionConfig, GpuStatsSectionMsg,
 };
 use crate::app::pages::PageUpdate;
